+++ device-bootloaders/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-6851]: Add support to reset ETSOC through PMIC
 - Added more clear debug info for PMIC error
 - [SW-14309] Extend Flash template to add PMIC FW img 
+- [SW-14479] Add API to change module frequency
 ### Changed
 - [SW-14084]: Fixing the CMA calculation to scale up and down the values correctly.
 - Support to read PMIC FW from PMIC directly versus hardcoded value